Consolidate animation classes into single package
