--- bigdoors-core/src/main/java/nl/pim16aap2/bigdoors/api/animatedblockhook/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains classes and interfaces for animated block hooks.
- */
-package nl.pim16aap2.bigdoors.api.animatedblockhook;
