+++ bigdoors-spigot/spigot-util/src/main/java/nl/pim16aap2/bigdoors/spigot/util/api/IAnimatedBlockHookSpigot.java
@@ -1,6 +1,6 @@
 package nl.pim16aap2.bigdoors.spigot.util.api;
 
-import nl.pim16aap2.bigdoors.api.animatedblockhook.IAnimatedBlockHook;
+import nl.pim16aap2.bigdoors.api.animatedblock.IAnimatedBlockHook;
 
 public interface IAnimatedBlockHookSpigot extends IAnimatedBlockHook<IAnimatedBlockSpigot>
 {
