+++ bigdoors-spigot/spigot-v1_15_R1/src/main/java/nl/pim16aap2/bigdoors/spigot/v1_15_R1/AnimatedBlockFactory_V1_15_R1.java
@@ -1,8 +1,8 @@
 package nl.pim16aap2.bigdoors.spigot.v1_15_R1;
 
 import nl.pim16aap2.bigdoors.api.IPLocation;
+import nl.pim16aap2.bigdoors.api.animatedblock.AnimationContext;
 import nl.pim16aap2.bigdoors.api.animatedblock.IAnimatedBlock;
-import nl.pim16aap2.bigdoors.api.animatedblockhook.AnimationContext;
 import nl.pim16aap2.bigdoors.api.factories.IAnimatedBlockFactory;
 import nl.pim16aap2.bigdoors.managers.AnimatedBlockHookManager;
 import nl.pim16aap2.bigdoors.spigot.util.SpigotAdapter;
