+++ bigdoors-spigot/spigot-v1_15_R1/src/main/java/nl/pim16aap2/bigdoors/spigot/v1_15_R1/CustomEntityFallingBlock_V1_15_R1.java
@@ -20,9 +20,9 @@ import net.minecraft.server.v1_15_R1.Vec3D;
 import net.minecraft.server.v1_15_R1.WorldServer;
 import nl.pim16aap2.bigdoors.api.IPLocation;
 import nl.pim16aap2.bigdoors.api.IPWorld;
-import nl.pim16aap2.bigdoors.api.animatedblockhook.AnimationContext;
-import nl.pim16aap2.bigdoors.api.animatedblockhook.IAnimatedBlockHook;
+import nl.pim16aap2.bigdoors.api.animatedblock.AnimationContext;
 import nl.pim16aap2.bigdoors.api.animatedblock.IAnimatedBlock;
+import nl.pim16aap2.bigdoors.api.animatedblock.IAnimatedBlockHook;
 import nl.pim16aap2.bigdoors.managers.AnimatedBlockHookManager;
 import nl.pim16aap2.bigdoors.spigot.util.SpigotAdapter;
 import nl.pim16aap2.bigdoors.spigot.util.api.IAnimatedBlockSpigot;
